moving sqlstore connection to a subpackage
