+++ pkg/infra/db/dbconn/dbconn_test.go
@@ -1,18 +1,14 @@
-package sqlstore
+package dbconn
 
 import (
-	"context"
 	"errors"
 	"net/url"
 	"testing"
-	"time"
-
-	"github.com/stretchr/testify/assert"
-	"github.com/stretchr/testify/require"
 
 	"github.com/grafana/grafana/pkg/services/featuremgmt"
-	"github.com/grafana/grafana/pkg/services/org"
 	"github.com/grafana/grafana/pkg/setting"
+	"github.com/stretchr/testify/assert"
+	"github.com/stretchr/testify/require"
 )
 
 type sqlStoreTest struct {
@@ -112,60 +108,12 @@ func TestIntegrationSQLConnectionString(t *testing.T) {
 	}
 	for _, testCase := range sqlStoreTestCases {
 		t.Run(testCase.name, func(t *testing.T) {
-			sqlstore := &SQLStore{}
-			sqlstore.Cfg = makeSQLStoreTestConfig(t, testCase)
-			connStr, err := sqlstore.buildConnectionString()
+			conn := &Conn{}
+			cfg := makeSQLStoreTestConfig(t, testCase)
+			connStr, err := conn.buildConnectionString(cfg)
 			require.Equal(t, testCase.err, err)
 
 			assert.Equal(t, testCase.expConnStr, connStr)
-		})
-	}
-}
-
-func TestIntegrationIsUniqueConstraintViolation(t *testing.T) {
-	store := InitTestDB(t)
-
-	testCases := []struct {
-		desc string
-		f    func(*testing.T, *DBSession) error
-	}{
-		{
-			desc: "successfully detect primary key violations",
-			f: func(t *testing.T, sess *DBSession) error {
-				// Attempt to insert org with provided ID (primary key) twice
-				now := time.Now()
-				org := org.Org{Name: "test org primary key violation", Created: now, Updated: now, ID: 42}
-				err := sess.InsertId(&org, store.Dialect)
-				require.NoError(t, err)
-
-				// Provide a different name to avoid unique constraint violation
-				org.Name = "test org 2"
-				return sess.InsertId(&org, store.Dialect)
-			},
-		},
-		{
-			desc: "successfully detect unique constrain violations",
-			f: func(t *testing.T, sess *DBSession) error {
-				// Attempt to insert org with reserved name
-				now := time.Now()
-				org := org.Org{Name: "test org unique constrain violation", Created: now, Updated: now, ID: 43}
-				err := sess.InsertId(&org, store.Dialect)
-				require.NoError(t, err)
-
-				// Provide a different ID to avoid primary key violation
-				org.ID = 44
-				return sess.InsertId(&org, store.Dialect)
-			},
-		},
-	}
-
-	for _, tc := range testCases {
-		t.Run(tc.desc, func(t *testing.T) {
-			err := store.WithDbSession(context.Background(), func(sess *DBSession) error {
-				return tc.f(t, sess)
-			})
-			require.Error(t, err)
-			assert.True(t, store.Dialect.IsUniqueConstraintViolation(err))
 		})
 	}
 }
