+++ pkg/infra/db/dbconn/xorm_logger.go
@@ -1,21 +1,21 @@
-package sqlstore
+package dbconn
 
 import (
 	"fmt"
 
-	"xorm.io/core"
-
 	glog "github.com/grafana/grafana/pkg/infra/log"
+
+	"xorm.io/core"
 )
 
-type XormLogger struct {
+type xormLogger struct {
 	grafanaLog glog.Logger
 	level      glog.Lvl
 	showSQL    bool
 }
 
-func NewXormLogger(level glog.Lvl, grafanaLog glog.Logger) *XormLogger {
-	return &XormLogger{
+func newXormLogger(level glog.Lvl, grafanaLog glog.Logger) *xormLogger {
+	return &xormLogger{
 		grafanaLog: grafanaLog,
 		level:      level,
 		showSQL:    true,
@@ -23,63 +23,63 @@ func NewXormLogger(level glog.Lvl, grafanaLog glog.Logger) *XormLogger {
 }
 
 // Error implement core.ILogger
-func (s *XormLogger) Error(v ...any) {
+func (s *xormLogger) Error(v ...any) {
 	if s.level <= glog.LvlError {
 		s.grafanaLog.Error(fmt.Sprint(v...))
 	}
 }
 
 // Errorf implement core.ILogger
-func (s *XormLogger) Errorf(format string, v ...any) {
+func (s *xormLogger) Errorf(format string, v ...any) {
 	if s.level <= glog.LvlError {
 		s.grafanaLog.Error(fmt.Sprintf(format, v...))
 	}
 }
 
 // Debug implement core.ILogger
-func (s *XormLogger) Debug(v ...any) {
+func (s *xormLogger) Debug(v ...any) {
 	if s.level <= glog.LvlDebug {
 		s.grafanaLog.Debug(fmt.Sprint(v...))
 	}
 }
 
 // Debugf implement core.ILogger
-func (s *XormLogger) Debugf(format string, v ...any) {
+func (s *xormLogger) Debugf(format string, v ...any) {
 	if s.level <= glog.LvlDebug {
 		s.grafanaLog.Debug(fmt.Sprintf(format, v...))
 	}
 }
 
 // Info implement core.ILogger
-func (s *XormLogger) Info(v ...any) {
+func (s *xormLogger) Info(v ...any) {
 	if s.level <= glog.LvlInfo {
 		s.grafanaLog.Info(fmt.Sprint(v...))
 	}
 }
 
 // Infof implement core.ILogger
-func (s *XormLogger) Infof(format string, v ...any) {
+func (s *xormLogger) Infof(format string, v ...any) {
 	if s.level <= glog.LvlInfo {
 		s.grafanaLog.Info(fmt.Sprintf(format, v...))
 	}
 }
 
 // Warn implement core.ILogger
-func (s *XormLogger) Warn(v ...any) {
+func (s *xormLogger) Warn(v ...any) {
 	if s.level <= glog.LvlWarn {
 		s.grafanaLog.Warn(fmt.Sprint(v...))
 	}
 }
 
 // Warnf implement core.ILogger
-func (s *XormLogger) Warnf(format string, v ...any) {
+func (s *xormLogger) Warnf(format string, v ...any) {
 	if s.level <= glog.LvlWarn {
 		s.grafanaLog.Warn(fmt.Sprintf(format, v...))
 	}
 }
 
 // Level implement core.ILogger
-func (s *XormLogger) Level() core.LogLevel {
+func (s *xormLogger) Level() core.LogLevel {
 	switch s.level {
 	case glog.LvlError:
 		return core.LOG_ERR
@@ -95,11 +95,11 @@ func (s *XormLogger) Level() core.LogLevel {
 }
 
 // SetLevel implement core.ILogger
-func (s *XormLogger) SetLevel(l core.LogLevel) {
+func (s *xormLogger) SetLevel(l core.LogLevel) {
 }
 
 // ShowSQL implement core.ILogger
-func (s *XormLogger) ShowSQL(show ...bool) {
+func (s *xormLogger) ShowSQL(show ...bool) {
 	if len(show) == 0 {
 		s.showSQL = true
 		return
@@ -108,6 +108,6 @@ func (s *XormLogger) ShowSQL(show ...bool) {
 }
 
 // IsShowSQL implement core.ILogger
-func (s *XormLogger) IsShowSQL() bool {
+func (s *xormLogger) IsShowSQL() bool {
 	return s.showSQL
 }
